move extensions apiHooks into their own files
